+++ src/Rib.hs
@@ -2,23 +2,20 @@
 module Rib
   ( module Rib.App,
     module Rib.Shake,
-    module Rib.Server,
     Document,
     documentPath,
     documentVal,
     documentHtml,
     documentMeta,
     documentUrl,
-    MMark,
+    Markup (..),
     renderMarkdown,
-    Pandoc,
     renderPandoc,
   )
 where
 
 import Rib.App
 import Rib.Document
-import Rib.Markup.MMark (MMark, renderMarkdown)
-import Rib.Markup.Pandoc (Pandoc, renderPandoc)
-import Rib.Server
+import Rib.Markup.MMark (renderMarkdown)
+import Rib.Markup.Pandoc (renderPandoc)
 import Rib.Shake